+++ verifier/src/test/java/com/aws/jverify/verifier/tests/generics/PolymorphismWithoutBounds.java
@@ -5,7 +5,7 @@ import com.aws.jverify.testengine.JVerifyTest;
 
 import static com.aws.jverify.JVerify.*;
 
-@JVerifyTest(dafnyVerified = 6, dafnyErrors = 0)
+@JVerifyTest(dafnyVerified = 4, dafnyErrors = 0)
 public class PolymorphismWithoutBounds {
 
     public static <T> void objectIsTop(T value) {
@@ -20,9 +20,6 @@ public class PolymorphismWithoutBounds {
 
         GenericContainer<DummySuper> container2 = new GenericContainer<>(obj);
         check(obj == container2.getValue());
-
-        var container3 = new GenericContainer.NestedGenericContainer<>(obj);
-        check(obj == container3.getValue());
         
         var obj2 = PolymorphismWithoutBounds.<DummySuper>genericIdentity(obj);
         check(obj == obj2);
@@ -35,40 +32,20 @@ public class PolymorphismWithoutBounds {
     public static <T> T genericIdentity(T a) {
         return a;
     }
-}
 
-class DummySuper { }
-class Dummy extends DummySuper {}
+    static class DummySuper { }
+    static class Dummy extends DummySuper {}
 
-class GenericContainer<T> {
-    private T value;
-    NestedGenericContainer<Dummy> nestedGenericContainer;
-    NestedGenericContainer<T> nestedGenericContainerT;
+    static class GenericContainer<T> {
+        private T value;
 
-    public GenericContainer(T value) {
-        postcondition(this.value == value);
-        this.value = value;
-        var dummy = new Dummy();
-        nestedGenericContainer = new NestedGenericContainer<>(dummy);
-        nestedGenericContainerT = new NestedGenericContainer<>(value);
-    }
-
-    @Pure
-    public T getValue() {
-        reads(this);
-        return value;
-    }
-    
-    public static class NestedGenericContainer<U> {
-        private U value;
-
-        public NestedGenericContainer(U value) {
+        public GenericContainer(T value) {
             postcondition(this.value == value);
             this.value = value;
         }
 
         @Pure
-        public U getValue() {
+        public T getValue() {
             reads(this);
             return value;
         }
